requirements.txt
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
cbor2==5.9.0
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.7
cloudinary==1.44.1
constantly==23.10.4
cryptography==46.0.7
daphne==4.2.1
Django==6.0.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
mongoengine==0.29.3
msgpack==1.1.2
packaging==26.0
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
PyJWT==2.12.1
pymongo==4.16.0
pyOpenSSL==26.0.0
python-dotenv==1.2.2
requests==2.33.1
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
ujson==5.12.0
urllib3==2.6.3
zope.interface==8.3
